backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
httpx==0.27.0
PyMuPDF==1.24.0
Pillow==10.2.0
google-generativeai==0.4.1
numpy==1.26.4
websockets==12.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
firebase-admin==6.4.0
